requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
bandit==1.7.2
cffi==1.15.0
colorama==0.4.4
Django==4.0.2
django-extensions==3.1.5
djangorestframework==3.13.1
drf-nested-routers==0.93.4
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.26
mccabe==0.6.1
pbr==5.8.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pytz==2021.3
PyYAML==6.0
smmap==5.0.0
sqlparse==0.4.2
stevedore==3.5.0
tzdata==2021.5
Werkzeug==2.0.3
